requirements.txt
python-dotenv

streamlit

langchain
langchain-core
langgraph

# Embeddings (local, no API key)
sentence-transformers

# LLM providers
langchain-groq
langchain-google-genai

chromadb
langchain-chroma
faiss-cpu

pymupdf
sqlalchemy
